+++ wazuh/appserver/static/js/agent_summary.js
@@ -1,0 +1,297 @@
+// <![CDATA[
+// <![CDATA[
+//
+// LIBRARY REQUIREMENTS
+//
+// In the require function, we include the necessary libraries and modules for
+// the HTML dashboard. Then, we pass variable names for these libraries and
+// modules as function parameters, in order.
+// 
+// When you add libraries or modules, remember to retain this mapping order
+// between the library or module and its function parameter. You can do this by
+// adding to the end of these lists, as shown in the commented examples below.
+
+require([
+  "splunkjs/mvc",
+  "splunkjs/mvc/utils",
+  "splunkjs/mvc/tokenutils",
+  "underscore",
+  "jquery",
+  "splunkjs/mvc/simplexml",
+  "splunkjs/mvc/layoutview",
+  "splunkjs/mvc/simplexml/dashboardview",
+  "splunkjs/mvc/simplexml/dashboard/panelref",
+  "splunkjs/mvc/simplexml/element/chart",
+  "splunkjs/mvc/simplexml/element/event",
+  "splunkjs/mvc/simplexml/element/html",
+  "splunkjs/mvc/simplexml/element/list",
+  "splunkjs/mvc/simplexml/element/map",
+  "splunkjs/mvc/simplexml/element/single",
+  "splunkjs/mvc/simplexml/element/table",
+  "splunkjs/mvc/simplexml/element/visualization",
+  "splunkjs/mvc/simpleform/formutils",
+  "splunkjs/mvc/simplexml/eventhandler",
+  "splunkjs/mvc/simplexml/searcheventhandler",
+  "splunkjs/mvc/simpleform/input/dropdown",
+  "splunkjs/mvc/simpleform/input/radiogroup",
+  "splunkjs/mvc/simpleform/input/linklist",
+  "splunkjs/mvc/simpleform/input/multiselect",
+  "splunkjs/mvc/simpleform/input/checkboxgroup",
+  "splunkjs/mvc/simpleform/input/text",
+  "splunkjs/mvc/simpleform/input/timerange",
+  "splunkjs/mvc/simpleform/input/submit",
+  "splunkjs/mvc/searchmanager",
+  "splunkjs/mvc/savedsearchmanager",
+  "splunkjs/mvc/postprocessmanager",
+  "splunkjs/mvc/simplexml/urltokenmodel"
+  // Add comma-separated libraries and modules manually here, for example:
+  // ..."splunkjs/mvc/simplexml/urltokenmodel",
+  // "splunkjs/mvc/tokenforwarder"
+],
+  function (
+    mvc,
+    utils,
+    TokenUtils,
+    _,
+    $,
+    DashboardController,
+    LayoutView,
+    Dashboard,
+    PanelRef,
+    ChartElement,
+    EventElement,
+    HtmlElement,
+    ListElement,
+    MapElement,
+    SingleElement,
+    TableElement,
+    VisualizationElement,
+    FormUtils,
+    EventHandler,
+    SearchEventHandler,
+    DropdownInput,
+    RadioGroupInput,
+    LinkListInput,
+    MultiSelectInput,
+    CheckboxGroupInput,
+    TextInput,
+    TimeRangeInput,
+    SubmitButton,
+    SearchManager,
+    SavedSearchManager,
+    PostProcessManager,
+    UrlTokenModel
+
+    // Add comma-separated parameter names here, for example: 
+    // ...UrlTokenModel, 
+    // TokenForwarder
+  ) {
+
+    var pageLoading = true;
+
+
+    // 
+    // TOKENS
+    //
+
+    // Create token namespaces
+    var urlTokenModel = new UrlTokenModel();
+    mvc.Components.registerInstance('url', urlTokenModel);
+    var defaultTokenModel = mvc.Components.getInstance('default', { create: true });
+    var submittedTokenModel = mvc.Components.getInstance('submitted', { create: true });
+    var service = mvc.createService({ owner: "nobody" });
+
+    urlTokenModel.on('url:navigate', function () {
+      defaultTokenModel.set(urlTokenModel.toJSON());
+      if (!_.isEmpty(urlTokenModel.toJSON()) && !_.all(urlTokenModel.toJSON(), _.isUndefined)) {
+        submitTokens();
+      } else {
+        submittedTokenModel.clear();
+      }
+    });
+
+    // Initialize tokens
+    defaultTokenModel.set(urlTokenModel.toJSON());
+
+    function submitTokens() {
+      // Copy the contents of the defaultTokenModel to the submittedTokenModel and urlTokenModel
+      FormUtils.submitForm({ replaceState: pageLoading });
+    }
+
+    function setToken(name, value) {
+      defaultTokenModel.set(name, value);
+      submittedTokenModel.set(name, value);
+    }
+
+    function unsetToken(name) {
+      defaultTokenModel.unset(name);
+      submittedTokenModel.unset(name);
+    }
+
+    $(document).ready(function () {
+      service.request(
+        "storage/collections/data/credentials/",
+        "GET",
+        null,
+        null,
+        null,
+        { "Content-Type": "application/json" }, null
+      ).done(function (data) {
+        var parsedData = JSON.parse(data);
+        console.log(parsedData)
+        console.log('BASEIP', JSON.parse(data)[0].baseip);
+        setToken('baseip', parsedData[0].baseip);
+        setToken('baseport', parsedData[0].baseport);
+        setToken('ipapi', parsedData[0].ipapi);
+        setToken('portapi', parsedData[0].portapi);
+        setToken('userapi', parsedData[0].userapi);
+        setToken('passwordapi', parsedData[0].passapi);
+        setToken("loadedtokens", "true");
+      });
+    })
+
+    //
+    // SEARCH MANAGERS
+    //
+
+
+    var search1 = new SearchManager({
+      "id": "search1",
+      "cancelOnUnload": true,
+      "sample_ratio": 1,
+      "earliest_time": "-60m@m",
+      "status_buckets": 0,
+      "search": "| getagentscheck $baseip$ $baseport$ $ipapi$ $portapi$ $userapi$ $passwordapi$ |table id, ip, name, status, os-platform, os-uname, os-name, os-arch, os-version, dateAdd, lastKeepAlive, last_rootcheck, last_syscheck, version | dedup id | sort - id | rename os-platform as \"Platform\", os-uname as \"OS Info\", os-name as \"OS name\", os-arch as \"Arch\", os-version as \"OS Version\", dateAdd as \"Registered date\", lastKeepAlive as \"Last KeepAlive\", last_rootcheck as \"Last Rootcheck\", last_syscheck as \"Last Syscheck\", version as \"Agent version\" | fillnull value=\"N/A\" | search id=*$valueId$* name = *$valueName$* ip = *$valueIp$* \"Agent version\" = *$valueVersion$*",
+      "latest_time": "now",
+      "app": utils.getCurrentApp(),
+      "auto_cancel": 90,
+      "preview": true,
+      "tokenDependencies": {
+      },
+      "runWhenTimeIsUndefined": false
+    }, { tokens: true, tokenNamespace: "submitted" });
+
+
+    //
+    // SPLUNK LAYOUT
+    //
+
+    $('header').remove();
+    new LayoutView({ "hideFooter": false, "hideSplunkBar": false, "hideAppBar": false, "hideChrome": false })
+      .render()
+      .getContainerElement()
+      .appendChild($('.dashboard-body')[0]);
+
+    //
+    // DASHBOARD EDITOR
+    //
+
+    new Dashboard({
+      id: 'dashboard',
+      el: $('.dashboard-body'),
+      showTitle: true,
+      editable: true
+    }, { tokens: true }).render();
+
+
+    //
+    // VIEWS: VISUALIZATION ELEMENTS
+    //
+
+    var element1 = new TableElement({
+      "id": "element1",
+      "count": 100,
+      "dataOverlayMode": "none",
+      "drilldown": "cell",
+      "percentagesRow": "false",
+      "refresh.display": "progressbar",
+      "rowNumbers": "false",
+      "totalsRow": "false",
+      "wrap": "true",
+      "managerid": "search1",
+      "el": $('#element1')
+    }, { tokens: true, tokenNamespace: "submitted" }).render();
+
+
+    //
+    // VIEWS: FORM INPUTS
+    //
+
+    var input1 = new TextInput({
+      "id": "input1",
+      "initialValue": "*",
+      "searchWhenChanged": true,
+      "default": "*",
+      "value": "$form.valueId$",
+      "el": $('#input1')
+    }, { tokens: true }).render();
+
+    input1.on("change", function (newValue) {
+      FormUtils.handleValueChange(input1);
+    });
+
+    var input2 = new TextInput({
+      "id": "input2",
+      "initialValue": "*",
+      "searchWhenChanged": true,
+      "default": "*",
+      "value": "$form.valueName$",
+      "el": $('#input2')
+    }, { tokens: true }).render();
+
+    input2.on("change", function (newValue) {
+      FormUtils.handleValueChange(input2);
+    });
+
+    var input3 = new TextInput({
+      "id": "input3",
+      "initialValue": "*",
+      "searchWhenChanged": true,
+      "default": "*",
+      "value": "$form.valueIp$",
+      "el": $('#input3')
+    }, { tokens: true }).render();
+
+    input3.on("change", function (newValue) {
+      FormUtils.handleValueChange(input3);
+    });
+
+    var input4 = new TextInput({
+      "id": "input4",
+      "initialValue": "*",
+      "searchWhenChanged": true,
+      "default": "*",
+      "value": "$form.valueVersion$",
+      "el": $('#input4')
+    }, { tokens: true }).render();
+
+    input4.on("change", function (newValue) {
+      FormUtils.handleValueChange(input4);
+    });
+
+    DashboardController.onReady(function () {
+      if (!submittedTokenModel.has('earliest') && !submittedTokenModel.has('latest')) {
+        submittedTokenModel.set({ earliest: '0', latest: '' });
+      }
+    });
+
+    // Initialize time tokens to default
+    if (!defaultTokenModel.has('earliest') && !defaultTokenModel.has('latest')) {
+      defaultTokenModel.set({ earliest: '0', latest: '' });
+    }
+
+    if (!_.isEmpty(urlTokenModel.toJSON())) {
+      submitTokens();
+    }
+
+
+    //
+    // DASHBOARD READY
+    //
+
+    DashboardController.ready();
+    pageLoading = false;
+
+  }
+);
+// ]]>